pyproject.toml
[tool.poetry]
name = "notificationservice"
version = "0.1.0"
description = ""
authors = ["Kamran <kamrankek48@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.1"
djangochannelsrestframework = "^1.2.0"
djangorestframework = "^3.15.2"
daphne = "^4.1.2"
requests = "^2.32.3"
websockets = "^13.1"
aiohttp = "^3.10.9"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
